# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SwissPipe API
  description: |
    SwissPipe is a powerful workflow automation platform that processes data through DAG-based workflows with JavaScript integration.

    ## Features
    - **Workflow Engine**: DAG-based workflow processing with Trigger, Condition, Transformer, and App nodes
    - **JavaScript Runtime**: Safe script execution for transformers and conditions using QuickJS
    - **Async Execution**: Background job processing with worker pools and cleanup services
    - **Multi-Database Support**: SQLite (development) and PostgreSQL (production) support
    - **Authentication**: Basic Auth for admin APIs, UUID-based for workflow ingestion
    - **AI Integration**: Anthropic Claude integration for code and workflow generation

    ## Authentication
    - **Admin APIs** (`/api/admin/v1/*`): Require Basic Authentication (username/password)
    - **Workflow Ingestion** (`/api/v1/*`): Use UUID-based workflow authentication
    - **Google OAuth**: Available for UI authentication when configured

    ## Data Flow
    1. **Ingestion**: Data enters through workflow trigger endpoints
    2. **Processing**: Workflows execute as background jobs through the worker pool
    3. **Execution**: Each workflow step is tracked with input/output data
    4. **Monitoring**: Real-time execution status and detailed step information available
  version: 0.1.0
  contact:
    name: SwissPipe Team
    url: https://github.com/prabhatsharma/swisspipe
  license:
...
servers:
...
security:
  - basicAuth: []
  - workflowAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication for admin endpoints (username/password)
    workflowAuth:
      type: apiKey
      in: path
      name: workflow_id
      description: UUID-based workflow authentication
...
paths:
...
  /api/admin/v1/executions/stats:
    get:
      tags: [Admin - Monitoring]
      summary: Get Worker Pool Statistics
      description: Get worker pool and system statistics
      security:
        - basicAuth: []
      responses:
        '200':
          description: Worker pool statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  worker_pool:
                    type: object
                    description: Worker pool metrics
                  system:
                    type: object
                    properties:
                      timestamp:
                        type: integer
                        format: int64
                      version:
                        type: string
                      build_profile:
                        type: string
                        enum: [debug, release]
                  health:
                    type: string
                    example: healthy
  /api/admin/v1/executions/cleanup/stats:
    get:
      tags: [Admin - Monitoring]
      summary: Get Cleanup Statistics
      description: Get execution cleanup service statistics
      security:
        - basicAuth: []
      responses:
        '200':
          description: Cleanup statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  cleanup_stats:
                    type: object
                    description: Cleanup metrics
                  config:
                    type: object
                    properties:
                      retention_count:
                        type: integer
                      cleanup_interval_minutes:
                        type: integer
                  timestamp:
                    type: integer
                    format: int64

  # Script Execution API
...